backend/requirements.txt
Django
psycopg2-binary
# firebase-admin
djangorestframework
markdown
django-filter
httpie